implemented base tests '

+++ src/main/java/com/bjuan/tallerpruebas/dao/CartDAO.java
@@ -30,7 +30,7 @@ public class CartDAO {
         entityManager.remove(sci);
     }
 
-    public Shoppingcartitem findById(long sci) {
+    public Shoppingcartitem findById(Integer sci) {
         return entityManager.find(Shoppingcartitem.class, sci);
     }
